--- src/csv.rs
@@ -1,3 +1,0 @@
-/*
-* csv parsing
-*/
