+++ src/data.rs
@@ -1,5 +1,5 @@
 /*
-* custom data parsing and wrangling
+* custom data parsing and wrangling. includes csv parsing
 */
 use common;
