--- src/dfj.rs
@@ -1,3 +1,0 @@
-/*
-* Dantzig–Fulkerson–Johnson
-*/
