+++ src/tests.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 pub mod tests {
     use crate::common;
+    use crate::geo;
 
     #[test]
     fn test_valid_calculate_haversine() {
@@ -8,7 +9,7 @@ pub mod tests {
         let dest = [0.0, 0.0];
         let unit = common::DistanceUnit::MI;
 
-        let result = crate::cacluate_haversine(origin, dest, &unit);
+        let result = geo::cacluate_haversine(origin, dest, &unit);
         assert_eq!(result, 0.0);
     }
 }
